backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
pillow
torch
torchvision